flops in ops
