+++ vllm/hpu/ops.py
@@ -5,6 +5,7 @@
 # LICENSE file in the root directory of this source tree.
 ###############################################################################
 import os
+from math import ceil
 from typing import Optional
 
 import habana_frameworks.torch as htorch
@@ -78,6 +79,17 @@ def paged_attention_v1(query,
     if query_heads != kv_heads:
         attn_weights = [a.flatten(1, 2) for a in attn_weights]
     attn_weights = sum(attn_weights)
+
+    # import pdb; pdb.set_trace()
+
+    flops = flops_counter(num_att_heads=query.shape[1],
+                            query_seq_len=query.shape[2],
+                            block_size=block_size,
+                            context_lens=context_lens,
+                            query_embedding_dim=query.shape[3],
+                            value_embedding_dim=key_cache.shape[3])
+    print("FLOPS: ", flops)
+    
     return attn_weights.squeeze(-2)
 
 
@@ -151,3 +163,13 @@ def prompt_attention(
         attn_weights = attn_weights.flatten(1, 2)
     attn_weights = attn_weights.transpose(1, 2)
     return attn_weights
+
+
+def flops_counter(num_att_heads, 
+                   query_seq_len, 
+                   block_size, 
+                   context_lens, 
+                   query_embedding_dim, 
+                   value_embedding_dim) -> float:
+    return sum([num_att_heads * query_seq_len * ceil(S_i / block_size) * block_size 
+                * 2 * (query_embedding_dim + value_embedding_dim) for S_i in context_lens])
